pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "dev-orchestrator"
version = "0.1.0"
description = "AI-powered development task orchestrator via Telegram"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Dev Task Orchestrator Team" }
]
keywords = ["ai", "automation", "telegram", "development", "orchestrator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
]

dependencies = [
    # Telegram
    "python-telegram-bot>=21.0",
    
    # AI Providers
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    
    # HTTP
    "httpx>=0.27.0",
    
    # Git
    "GitPython>=3.1.40",
    "PyGithub>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Async
    "anyio>=4.0.0",
    
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    
    # Linting & Formatting
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    
    # Pre-commit
    "pre-commit>=3.6.0",
    
    # Type stubs
    "types-requests>=2.31.0",
]

[project.scripts]
dev-orchestrator = "src.main:main"
orchestrator-cli = "src.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/dev-orchestrator"
Documentation = "https://github.com/your-org/dev-orchestrator#readme"
Repository = "https://github.com/your-org/dev-orchestrator"
Issues = "https://github.com/your-org/dev-orchestrator/issues"

# =============================================================================
# RUFF
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "RUF",    # Ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",     # Unused arguments in tests (fixtures)
    "PLR2004", # Magic values in tests OK
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# =============================================================================
# MYPY
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "telegram.*",
    "git.*",
    "github.*",
    "google.*",
]
ignore_missing_imports = true

# =============================================================================
# PYTEST
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-fail-under=85",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, external deps)",
    "slow: Tests that take > 1s",
    "ai: Tests that mock AI providers",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/__main__.py",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
